+++ src/Digdir.Domain.Dialogporten.Domain/Dialogs/Entities/Attachments/Attachment.cs
@@ -5,18 +5,6 @@ using Digdir.Library.Entity.Abstractions.Features.Aggregate;
 
 namespace Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Attachments;
 
-public sealed class DialogAttachment : Attachment
-{
-    public Guid DialogId { get; set; }
-    public DialogEntity Dialog { get; set; } = null!;
-}
-
-public sealed class TransmissionAttachment : Attachment
-{
-    public Guid TransmissionId { get; set; }
-    public DialogTransmission Transmission { get; set; } = null!;
-}
-
 public abstract class Attachment : IEntity
 {
     public Guid Id { get; set; }
