+++ src/Digdir.Domain.Dialogporten.Domain/Dialogs/Entities/Transmissions/DialogTransmission.cs
@@ -2,6 +2,7 @@ using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Activities;
 using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Actors;
 using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Attachments;
 using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Contents;
+using Digdir.Library.Entity.Abstractions.Features.Aggregate;
 using Digdir.Library.Entity.Abstractions.Features.Immutable;
 
 namespace Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Transmissions;
@@ -10,18 +11,22 @@ public class DialogTransmission : IImmutableEntity
 {
     public Guid Id { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
-
     public string? AuthorizationAttribute { get; set; }
-
-    public DialogTransmissionSenderActor Sender { get; set; } = null!;
-
     public Uri? ExtendedType { get; set; }
 
     // === Principal relationships ===
+    [AggregateChild]
     public List<TransmissionContent> Content { get; set; } = [];
+
+    [AggregateChild]
     public List<TransmissionAttachment> Attachments { get; set; } = [];
-    public List<DialogActivity> Activities { get; set; } = [];
+
+    [AggregateChild]
+    public DialogTransmissionSenderActor Sender { get; set; } = null!;
+
     public List<DialogTransmission> RelatedTransmissions { get; set; } = [];
+
+    public List<DialogActivity> Activities { get; set; } = [];
 
     // === Dependent relationships ===
     public Guid DialogId { get; set; }
@@ -32,4 +37,16 @@ public class DialogTransmission : IImmutableEntity
 
     public DialogTransmissionType.Values TypeId { get; set; }
     public DialogTransmissionType Type { get; set; } = null!;
+}
+
+public sealed class DialogTransmissionSenderActor : DialogActor
+{
+    public Guid TransmissionId { get; set; }
+    public DialogTransmission Transmission { get; set; } = null!;
+}
+
+public sealed class TransmissionAttachment : Attachment
+{
+    public Guid TransmissionId { get; set; }
+    public DialogTransmission Transmission { get; set; } = null!;
 }
